Make the neural network test for MNIST handwritten digits an example. Upload archive zips so others can run the example.
