+++ src/partitions.rs
@@ -1,4 +1,4 @@
-use std::{iter::StepBy, ops::RangeInclusive, thread};
+use std::{ops::RangeInclusive, thread};
 
 pub struct Partitioned<T> {
     pub partitioned: T,
